Update patch set 1
Patch Set 1:

Edvard, 

You might have gained a little support had the picture contained anything that was racially or sexually discriminating. As it happens it did not. Using genuine causes, which people truly believe in, in an attempt to censor material which is not to ones individual taste, only serves to dilute and discredit the cause. 

The argument of "objectifying" women is not valid. Glamour and porn models are typically well paid and enjoy their work. To say they are being exploited is not true - if anyone is being exploited, it is the men who pay money to look at such pictures. When the shoe is on the other foot I've never heard a complaint from anyone (men or women) about exploitation of men in similar pictures (eg Chippendales). 

People who say they are "offended" by such pictures are normally just drawing attention to themselves. How can anyone be offended by a picture like this? It doesn't make a personal attack on any group or individual! I don't like some of the CM wallpapers, but I don't proclaim to be offended by them. 

Your assertion that they are juvenile is also flawed. Fun pictures like these are enjoyed by all age groups, not just the young. CM is a fun and alternative community project that people do in their own time for free. I don't think there are any expectations for any level of professionalism of the output - only that it is of a high standard of quality, and that is only an expectation set by the hard working developers themselves.

Patch-set: 1