Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

An OCD nitpick.  Please fix the commit message.  After "New wallpapers form the CM Community" please put the rest of the commit message on a new line.  It would also be a good idea to specify that all of these wallpapers are CC licensed.

Patch-set: 1