Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

I agree with Kasper...

And there is a lot of wallpaper apps on GPlay. Why don't you put this app on Market? For ex: Ricardo's theme apps for ICS are on Market.

Sorry we can't accept 25MB for nothing.

Patch-set: 3