Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

The reason for the large file size is that the "drawable-sw600dp-nodpi" directory is included in all builds regardless of screen density.  For example, in a build for a xhdpi device (Galaxy Nexus) both drawable-xhdpi and drawable-sw600dp-nodpi are included in the build.

Patch-set: 3