Update patch set 3
Patch Set 3:

I didn't want to do this but I'll have to change some of the PNGs to jpg and maybe run the others through optimization to get the sizes down.  Basically the highest resolution in the old version was 960x800, and that's the lowest in the new ICS code.

As someone mentioned above, its the 1920x1280 ones that are probably causing the biggest inflation, I'll be sure to get those down especially in the next patchset.

Thanks.

Patch-set: 3