Update patch set 2
Patch Set 2:

I believe it's more of an issue of how many characters can be displayed in a single line. If that's the case, then "CyanogenMod" would still exceed the character limit and would still be split into 2 lines. I'll try to test this when I get some time.

Patch-set: 2