Update patch set 2
Patch Set 2:

Yeah, those images are the ones that lack resolution.

What device are you actually testing these on? What's the display panel's native resolution? I'd love to ship more pixels but I'm pretty sure we'll have to have images on the server for that. 3-4 meg image assets in the ROM aren't going to fly.

Patch-set: 2