Update patch set 1
Patch Set 1:

The idea, for whatever it was worth, was to partially replace the handful of Cid-based backgrounds we dumped in an earlier patch. We went from having a bunch of very obviously CM wallpapers to none.

Ideally we could include versions with and without artwork, but until we can move the backgrounds out of the build and onto a server I don't see that happening.

Patch-set: 1
Reviewer: Henry Mason <6135@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0