Update patch set 1
Patch Set 1:

If we are going to rename this, the correct name should be org.cyanogenmod.wallpapers, but is it not going this change upgrade problems from cm-10.2 to cm-11? I think it's just a minor problem, but if a user has CMWallpapers as its default wallpaper, the upgrade process will change it to default.

Patch-set: 1